refactor: separate field class
